i think this image workflow will need to be optimised
a way of doing this maybe to still have a poster and backdrop field (and image for person) and these are the primary images and get overriden when primary images are added, that way i don't have to do expensive includes. I need to test how expensive these include statements are. Especially for querying 30+ films from the database.